Creates useSearch hook
